refactor!: using new bee api
